🎨 コンポーネント化
