+++ src/components/Header.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Header = () => {
+  return (
+    <div className=" bg-blue-800">
+      <header className="container m-0 flex h-16 items-center">
+        <span className=" text-lg text-white">Quote Card</span>
+      </header>
+    </div>
+  );
+};
+
+export default Header;
